Consolidate home page into one component and re-use the logic so we don't have to use server rewrite/redirect, but also home page can query params can be used
